cdrom: Rename cdrom to better explain the difference between cdrom and CDVD
