+++ pcsx2/CDVD/Ps1CD.h
