+++ pcsx2/ps2/Iop/IopHwRead.cpp
@@ -18,7 +18,7 @@
 #include "IopHw_Internal.h"
 #include "Sif.h"
 #include "Sio.h"
-#include "CDVD/CdRom.h"
+#include "CDVD/Ps1CD.h"
 #include "FW.h"
 #include "SPU2/spu2.h"
 #include "DEV9/DEV9.h"
