+++ pcsx2/ps2/Iop/IopHwWrite.cpp
@@ -18,7 +18,7 @@
 #include "Sif.h"
 #include "Sio.h"
 #include "FW.h"
-#include "CDVD/CdRom.h"
+#include "CDVD/Ps1CD.h"
 #include "SPU2/spu2.h"
 #include "DEV9/DEV9.h"
 #include "USB/USB.h"
